+++ client-cli/src/main/java/org/eclipse/dataspaceconnector/identityhub/cli/IdentityHubCli.java
@@ -18,8 +18,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import okhttp3.OkHttpClient;
 import org.eclipse.dataspaceconnector.identityhub.client.IdentityHubClient;
 import org.eclipse.dataspaceconnector.identityhub.client.IdentityHubClientImpl;
-import org.eclipse.dataspaceconnector.identityhub.client.VerifiableCredentialsJwtServiceImpl;
-import org.eclipse.dataspaceconnector.identityhub.credentials.VerifiableCredentialsJwtService;
+import org.eclipse.dataspaceconnector.identityhub.client.VerifiableCredentialsJwtMapperImpl;
+import org.eclipse.dataspaceconnector.identityhub.credentials.VerifiableCredentialsJwtMapper;
 import org.eclipse.dataspaceconnector.spi.monitor.ConsoleMonitor;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
@@ -35,7 +35,7 @@ public class IdentityHubCli {
 
     IdentityHubClient identityHubClient;
 
-    VerifiableCredentialsJwtService verifiableCredentialsJwtService;
+    VerifiableCredentialsJwtMapper verifiableCredentialsJwtMapper;
 
     public static void main(String... args) {
         CommandLine commandLine = getCommandLine();
@@ -58,6 +58,6 @@ public class IdentityHubCli {
         var objectMapper = new ObjectMapper();
         var monitor = new ConsoleMonitor();
         this.identityHubClient = new IdentityHubClientImpl(okHttpClient, objectMapper, monitor);
-        this.verifiableCredentialsJwtService = new VerifiableCredentialsJwtServiceImpl(objectMapper);
+        this.verifiableCredentialsJwtMapper = new VerifiableCredentialsJwtMapperImpl(objectMapper);
     }
 }
