+++ spi/identity-hub-spi/src/main/java/org/eclipse/dataspaceconnector/identityhub/credentials/VerifiableCredentialsJwtMapper.java
@@ -25,7 +25,7 @@ import java.util.Map;
 /**
  * Service with operations for manipulation of VerifiableCredentials in JWT format.
  */
-public interface VerifiableCredentialsJwtService {
+public interface VerifiableCredentialsJwtMapper {
 
     String VERIFIABLE_CREDENTIALS_KEY = "vc";
